deleting users implemented

+++ frontend/src/app/components/edit-user/edit-user.component.ts
@@ -115,15 +115,18 @@ export class EditUserComponent implements OnInit {
   }
 
   deleteUser() {
-   this.userService.deleteUser(this.user).subscribe({
-      next: () => {
-        window.alert('Successfully deleted the User');
-        this.authService.logoutUser();
-      },
-      error: (error) => {
-        window.alert('Error during deleting User: ' + error.error.message);
-      }
-    });
+    if (window.confirm('Do you really want to delete your account?')) {
+      this.userService.deleteUser(this.user).subscribe({
+        next: () => {
+          this.authService.logoutUser();
+          this.router.navigate(['/login']);
+          window.alert('Successfully deleted the User');
+        },
+        error: (error) => {
+          window.alert('Error during deleting User: ' + error.error.message);
+        }
+      });
+    }
   }
 
   updateUser() {
